requirements.txt
# Dataset Downloader Dependencies
kaggle>=1.5.16
datasets>=2.14.0
huggingface_hub>=0.16.0
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.0.0

# Optional: For better GitHub downloads
gitpython>=3.1.40

# New dependencies for unstructured data sources
praw>=7.7.1  # Reddit API
feedparser>=6.0.10  # RSS feeds
beautifulsoup4>=4.12.2  # HTML parsing
newspaper3k>=0.2.8  # Article extraction
lxml>=4.9.3  # XML/HTML parser
python-dotenv>=1.0.0  # Environment variables

# RAG & Vector Database
sentence-transformers>=2.2.0  # FREE embeddings
faiss-cpu>=1.7.4  # Vector database (local, free - fallback)
weaviate-client[agents]>=4.0.0  # Weaviate with QueryAgent support
langchain>=0.1.0  # RAG framework
langchain-openai>=0.0.5  # OpenAI integration for LangChain
langchain-community>=0.0.20  # Community integrations

# LLM APIs
openai>=1.0.0  # OpenAI API

# Data Processing
numpy>=1.24.0,<2.0  # Pin to NumPy 1.x for compatibility with PyTorch/sentence-transformers
scikit-learn>=1.3.0  # For similarity search

# Chat Interface
streamlit>=1.28.0  # Web UI
streamlit-chat>=0.1.1  # Chat components

# Utilities
tiktoken>=0.5.0  # Token counting
ddgs>=0.1.0  # Web search (free, no API key needed) - renamed from duckduckgo-search

# Export Dependencies
reportlab>=4.0.0  # PDF export
python-pptx>=0.6.21  # PowerPoint export
Pillow>=10.0.0  # Image processing for PowerPoint

# Chart generation (for Gamma.ai features)
matplotlib>=3.7.0  # For charts
plotly>=5.17.0     # For interactive charts (optional)

# Image processing enhancements
Pillow>=10.0.0     # Already have

# Additional utilities for new features
python-dateutil>=2.8.2  # Date parsing